requirements.txt
asn1crypto==0.24.0
backports.ssl-match-hostname==3.5.0.1
blessings==1.7
bpython==0.17.1.post35
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
configobj==4.7.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
curtsies==0.3.0
decorator==3.4.0
Django==1.11.13
django-crispy-forms==1.7.2
django-encrypted-model-fields==0.5.5
django-filter==1.1.0
django-guardian==1.4.9
djangorestframework==3.8.2
enum34==1.1.6
greenlet==0.4.13
httpie==0.9.9
idna==2.7
iniparse==0.4
ipaddress==1.0.16
itypes==1.1.0
Jinja2==2.10
jsonschema==2.5.1
kitchen==1.1.1
langtable==0.0.31
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==1.2.0
meld3==0.6.10
mistune==0.8.3
MySQL-python==1.2.5
ndg-httpsclient==0.5.0
nose==1.3.7
numpy==1.7.1
path.py==5.2
perf==0.1
pexpect==2.3
Pillow==2.0.0
pyasn1==0.4.3
pycparser==2.18
pycurl==7.19.0
Pygments==2.2.0
pygobject==3.22.0
pygpgme==0.3
pyliblzma==0.5.3
pyOpenSSL==18.0.0
pyparsing==1.5.6
python-augeas==0.5.0
python-dateutil==1.5
python-dmidecode==3.10.13
python-linux-procfs==0.4.9
pytz==2018.4
pyudev==0.15
pyxattr==0.5.1
PyYAML==3.10
pyzmq==14.3.1
repoze.lru==0.4
requests==2.19.1
schedutils==0.4
simplegeneric==0.8
six==1.9.0
slip==0.4.0
slip.dbus==0.4.0
supervisor==3.1.4
typing==3.6.4
uritemplate==3.0.0
urlgrabber==3.10
urllib3==1.23
urwid==1.1.1
uWSGI==2.0.18
wcwidth==0.1.7
yum-langpacks==0.4.2
yum-metadata-parser==1.1.4
